Only update the selected value during avatarUpdate when the value is not null
The PATCH to update avatars doesn't receive the email, so we can't decide if that's the selected avatar for the e-mail we are working with.

In the AvatarStorage.updateAvatar method, we should use the value from the updatedAvatar only if it's not null. Otherwise, we should keep the stored avatar's selected value.